實作前台功能 - 將前後台控制器分離並重構部分方法、 將購物、追蹤等功能與會員系統串連

+++ src/main/java/com/howard/rentalmytrack/controller/FrontendRentalMyTrackController.java
@@ -3,12 +3,15 @@ package com.howard.rentalmytrack.controller;
 import com.howard.rentalmytrack.dto.AddTrack;
 import com.howard.rentalmytrack.entity.RentalMyTrack;
 import com.howard.rentalmytrack.service.impl.RentalMyTrackServiceImpl;
+import com.roger.member.entity.Member;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpSession;
 import java.sql.Timestamp;
 import java.util.List;
 
@@ -24,20 +27,31 @@ public class FrontendRentalMyTrackController {
         return rentalMyTrackService.getAll();
     }
 
+    String toLogin = "/frontend/member/loginMember";
+
     @GetMapping("/toRentalMyTrack")
-    public String toRentalMyTrack() {
+    public String toRentalMyTrack(HttpSession session, ModelMap model) {
+        Member member = (Member) session.getAttribute("loginsuccess");
+        if (member == null) {
+            return "redirect:" + toLogin;
+        }
+        model.addAttribute("member", member);
         return "frontend/rental/rentalMyTrack";
     }
 
     @PostMapping("/addTrack")
-    public ResponseEntity<?> addTrack(@RequestBody AddTrack addTrack) {
+    public ResponseEntity<?> addTrack(@RequestBody AddTrack addTrack, HttpSession session) {
+        Member member = (Member) session.getAttribute("loginsuccess");
+        if (member == null) {
+            return ResponseEntity.status(HttpStatus.OK).body(toLogin);
+        }
         // 若已經存在，一樣返回加入成功
-        RentalMyTrack testTrack = rentalMyTrackService.findById(addTrack.getRentalNo(), addTrack.getMemNo());
+        RentalMyTrack testTrack = rentalMyTrackService.findById(addTrack.getRentalNo(), member.getMemNo());
         if (testTrack != null) {
             return ResponseEntity.status(HttpStatus.CREATED).body(testTrack.getRental().getRentalName());
         }
         Timestamp expRentalDate = new Timestamp(System.currentTimeMillis());
-        RentalMyTrack.CompositeTrack compositeTrack = new RentalMyTrack.CompositeTrack(addTrack.getRentalNo(), addTrack.getMemNo());
+        RentalMyTrack.CompositeTrack compositeTrack = new RentalMyTrack.CompositeTrack(addTrack.getRentalNo(), member.getMemNo());
         RentalMyTrack rentalMyTrack = new RentalMyTrack(compositeTrack, expRentalDate);
         rentalMyTrackService.insert(rentalMyTrack);
         //
@@ -45,7 +59,7 @@ public class FrontendRentalMyTrackController {
         String rentalName = null;
         int tries = 10;
         while (tries-- > 0) {
-            trackForReturn = rentalMyTrackService.findById(addTrack.getRentalNo(), addTrack.getMemNo());
+            trackForReturn = rentalMyTrackService.findById(addTrack.getRentalNo(), member.getMemNo());
             if (trackForReturn.getRental() != null && trackForReturn.getRental() != null) {
                 rentalName = trackForReturn.getRental().getRentalName();
                 break;
@@ -63,10 +77,13 @@ public class FrontendRentalMyTrackController {
 
     // 刪除
     @DeleteMapping("/deleteTrack")
-    public ResponseEntity<?> deleteTrack(@RequestParam Integer rentalNo,
-                                         @RequestParam Integer memNo) {
+    public ResponseEntity<?> deleteTrack(@RequestParam Integer rentalNo, HttpSession session) {
 
-        RentalMyTrack rentalMyTrack = rentalMyTrackService.findById(rentalNo, memNo);
+        Member member = (Member) session.getAttribute("loginsuccess");
+        if (member == null) {
+            return ResponseEntity.status(HttpStatus.OK).body(toLogin);
+        }
+        RentalMyTrack rentalMyTrack = rentalMyTrackService.findById(rentalNo, member.getMemNo());
         if (rentalMyTrack != null) {
             rentalMyTrackService.delete(rentalMyTrack);
             return ResponseEntity.status(HttpStatus.OK).body("ok");
